Refactor into installable package
