+++ .gitignore
@@ -2,3 +2,4 @@
 config.py
 *.dat
 *.db
+rabbit.egg-info
