+++ README.md
@@ -1,0 +1,3 @@
+# rabbit
+
+A chat assistant for Stack Overflow room owners.
