+++ browser/Extended_User_Notes.user.js
