+++ rabbit/main.py
@@ -3,7 +3,7 @@ if sys.version_info < (3,0,0):
     print("Please run me in Python 3.")
     sys.exit(0)
 
-from stackoverflowchatsession import StackOverflowChatSession
+from .stackoverflowchatsession import StackOverflowChatSession
 import config
 
 import asyncio
@@ -11,7 +11,7 @@ import json
 import html
 import random
 from queue import Queue
-from dbmodel import User, db_session
+from .dbmodel import User, get_session
 
 PYTHON_ROOM_ID = 6
 PERSONAL_SANDBOX_ROOM_ID = 118024
@@ -103,9 +103,9 @@ class Rabbit(StackOverflowChatSession):
                 elif event_type == 15: #account level changed
                     if "created" in event["content"]: #kicked. (this may also catch other kinds of account level changed events, but they seem rare enough, and the side effects are innocuous enough, that I can debug them as I encounter them.
                         #record event.
-                        user = User.get_or_create(db_session, event["user_id"])
+                        user = User.get_or_create(get_session(), event["user_id"])
                         user.kick_count += 1
-                        db_session.commit()
+                        get_session().commit()
 
                         #now post a picture of a bunny.
                         bunny_url = random.choice(config.kick_reply_images)
@@ -177,8 +177,10 @@ def create_admin_window(message_queue):
 
     root.mainloop()
 
-message_queue = Queue()
-t = threading.Thread(target=create_admin_window, args=(message_queue,))
-t.start()
 
-create_and_run_chat_session(message_queue)
+def main():
+    message_queue = Queue()
+    t = threading.Thread(target=create_admin_window, args=(message_queue,))
+    t.start()
+
+    create_and_run_chat_session(message_queue)
